Update reCAPTCHA integration in login form
The implementation of the reCAPTCHA service in the login form was updated to provide better validation. ReCaptcha class was refactored for efficient validation checks. Changes in the code involved utilizing input form of reCAPTCHA in login.blade.php and modifying its validator in ReCaptcha.php. A new validator was also added to LoginController.php.